Move wait_for_signals to private module and deprecate distributed.cli.utils
